Accept / process orders is working but needs error testiing
I need to get transaction working but shouldn't be hard. UI accepts an
order -> warehouse createsOrder on 1st loop and then uses continueOrder
on additional products to the same order.  Process order is a private
function in warehouse.  I had to change client, invoice, invoiceentry,
order, product, and waitlistentry an addition to UI and warehouse.

+++ InvoiceEntry.java
@@ -3,16 +3,16 @@ import java.util.*;
 import java.io.*;
 public class InvoiceEntry implements Serializable {
     private String productId;
-    private float quantity;
+    private int quantity;
     private float price;
 
-    public InvoiceEntry(String productId, float quantity, float price) {
+    public InvoiceEntry(String productId, int quantity, float price) {
         this.productId = productId;
         this.quantity = quantity;
         this.price = price;
     }
 
-    public float getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
